move scenario setup to common and set scenarios in cfg
