+++ remind_mfa/common/common_cfg.py
@@ -91,7 +91,7 @@ class GeneralCfg(RemindMFABaseModel):
     madrat_output_path: str
     input_data_path: str
     input_data_version: str
-    force_extract: bool
+    scenario: dict
     customization: ModelCustomization
     visualization: VisualizationCfg
     output_path: str
